shop/requirements.txt
asgiref==3.3.4
astroid==2.12.11
Authlib==0.15.5
brotlipy==0.7.0
certifi==2021.10.8
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
colorama==0.4.5
coverage==6.5.0
cryptography==36.0.1
dill==0.3.5.1
Django==3.2.4
factory-boy==3.2.1
Faker==15.0.0
flake8==4.0.1
fpdf==1.7.2
idna==3.4
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
platformdirs==2.5.2
plotly==5.10.0
psycopg2==2.9.4
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
pylint==2.15.4
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2021.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.1
tenacity==8.1.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.5
typing_extensions==4.4.0
urllib3==1.26.12
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.14.1
